Fixes a couple more typing issues
