+++ src/sentry/integrations/github/integration.py
@@ -39,7 +39,6 @@ from sentry.integrations.models.organization_integration import OrganizationInte
 from sentry.integrations.pipeline import IntegrationPipeline
 from sentry.integrations.referrer_ids import GITHUB_OPEN_PR_BOT_REFERRER, GITHUB_PR_BOT_REFERRER
 from sentry.integrations.services.integration import integration_service
-from sentry.integrations.services.integration.model import RpcIntegration
 from sentry.integrations.services.repository import RpcRepository, repository_service
 from sentry.integrations.source_code_management.commit_context import (
     OPEN_PR_MAX_FILES_CHANGED,
@@ -297,18 +296,6 @@ class GitHubIntegration(
             permissions=self.model.metadata.get("permissions"),
             expires_at=expiration_time,
         )
-
-    def _update_integration_model(self) -> None:
-        updated_model: RpcIntegration | Integration | None = None
-        if isinstance(self.model, Integration):
-            updated_model = Integration.objects.get(id=self.model.id)
-        else:
-            updated_model = integration_service.get_integration(self.model.id)
-
-        if updated_model is None:
-            raise IntegrationError("Integration model not found")
-
-        self.model = updated_model
 
     # IntegrationInstallation methods
